split customised_instruments into files
