+++ qdev_wrappers/customised_instruments/Keysight_34465A_T10.py
@@ -1,0 +1,25 @@
+from qcodes.instrument_drivers.Keysight.Keysight_34465A import Keysight_34465A
+
+
+# Subclass the DMM
+class Keysight_34465A_T10(Keysight_34465A):
+    """
+    A Keysight DMM with an added I-V converter
+    """
+
+    def __init__(self, name, address, **kwargs):
+        super().__init__(name, address, **kwargs)
+
+        self.iv_conv = 1
+
+        self.add_parameter('ivconv',
+                           label='Current',
+                           unit='pA',
+                           get_cmd=self._get_current,
+                           set_cmd=None)
+
+    def _get_current(self):
+        """
+        get_cmd for dmm readout of IV_TAMP parameter
+        """
+        return self.volt() / self.iv_conv * 1E12
